Update 1.1
Added digital circuit to a new sheet in "VoltageRegulation" file, it appears we need the entire schematic in one sheet however. Digital circuit and the arduino were also added. In another folder, I will share the library used for the arduino nano footprint. It should be noted that the arduino library file found in the "LibUpdates" folder MUST be included in your own EAGLE lib folder in order to use the arduino in the schematic.