Update 1.3
All major parts have now been included in the schematic, only adjustments need to be made now. I know some resistor values have been implemented with series or parallel combinations in order use stockroom parts, and this is not accounted for. Values still need to be added to parts for the bill of materials as well as the silkscreening process. I've tried laying some traces as well in the board layout, but the config i've been trying out seems bad and needs an entire rework. We are at the point that board layout and tracing can commence.